requirements.txt
flask>=3.0
flask-sqlalchemy>=3.1
flask-migrate>=4.0
twilio>=9.0
gunicorn>=21.0
psycopg2-binary>=2.9
python-dotenv>=1.0
